+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/services/DataSourceRepository.scala
@@ -6,7 +6,7 @@ import com.google.inject.name.Named
 import com.scalableminds.webknossos.datastore.models.datasource.inbox.InboxDataSource
 import com.scalableminds.webknossos.datastore.models.datasource.{DataSource, DataSourceId}
 import com.scalableminds.webknossos.datastore.storage.TemporaryStore
-import com.scalableminds.util.tools.{Fox, FoxImplicits, TimeLogger}
+import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import com.typesafe.scalalogging.LazyLogging
 
 import scala.concurrent.ExecutionContext.Implicits.global
@@ -33,6 +33,6 @@ class DataSourceRepository @Inject()(
       _ <- Fox.successful(())
       _ = removeAll
       _ = dataSources.foreach(dataSource => insert(dataSource.id, dataSource))
-      _ <- TimeLogger.logTimeF("report datasources to wk", logger)(webKnossosServer.reportDataSources(dataSources))
+      _ <- webKnossosServer.reportDataSources(dataSources)
     } yield ()
 }