+++ CHANGELOG.md
@@ -16,7 +16,7 @@ For upgrade instructions, please check the [migration guide](MIGRATIONS.md).
 - The dataset table in the dashboard is now horizontally scrollable if the window is not wide enough. Additionally, clicking on the name of a dataset in the table opens of the dataset in view mode. [#4136](https://github.com/scalableminds/webknossos/pull/4136)
 
 ### Changed
-- 
+- Reactivated and renamed the "Quality" setting to "Hardware Utilization". Using a higher value will render data in higher quality, but puts more stress on the user's hardware and bandwidth. [#4142](https://github.com/scalableminds/webknossos/pull/4142)
 
 ### Fixed
 - Fixed that team managers couldn't view time tracking details of other users anymore. [#4125](https://github.com/scalableminds/webknossos/pull/4125)